added menu screen and additional cleanup
pause screen added, as well as menu screen.  added code to set
time.timescale to zero when escape key was pressed. added physics
material to player and terrain so that the player didn't stick to
terrain when falling off of the sides.  game is for the most part
complete, but still need to update so that the score doesn't carry over
when you try to restart the game.